main: fix shit

+++ agnostic/install-liver-zshtheme.sh
@@ -6,5 +6,11 @@ git clone "https://$repo_url" "$dest_path"
 
 cp "$dest_path/liver.zsh-theme" "$HOME/.oh-my-zsh/themes/liver.zsh-theme"
 
-rm -rf "$dest_path"
-
+echo -n "Do you want to delete the local git repository? (y/n)"
+read reply
+if [ $reply == "y" ]; then
+    echo "Removing '$dest_path'"
+    rm -rf "$dest_path"
+else
+    echo "Not removing '$dest_path'"
+fi
